Fixed codeclimate issues
