+++ app/controllers/concerns/epp_requestable.rb
@@ -9,8 +9,8 @@ module EppRequestable
     authorize! :create, Epp::Server
     result = server.request(request_params[:payload])
     render_success(data: { xml: result.force_encoding('UTF-8') })
-  rescue StandardError
-    handle_non_epp_errors(nil, I18n.t('errors.messages.epp_conn_error'))
+  rescue StandardError => e
+    handle_non_epp_errors(nil, e.message.presence || I18n.t('errors.messages.epp_conn_error'))
   end
 
   private
